loaders and messages added

+++ client-app/src/app/users-list/users-list.component.ts
@@ -2,6 +2,8 @@ import { Component, OnInit } from '@angular/core';
 import { UserService } from '../services/user.service';
 import { Router } from '@angular/router';
 
+declare var swal: any;
+
 @Component({
   selector: 'app-users-list',
   templateUrl: './users-list.component.html',
@@ -10,6 +12,7 @@ import { Router } from '@angular/router';
 export class UsersListComponent implements OnInit {
 
   Users = [];
+  loader:boolean=true;
   constructor(private userService: UserService, private router: Router) { }
 
   ngOnInit(): void {
@@ -19,9 +22,11 @@ export class UsersListComponent implements OnInit {
   getUsers() {
     this.userService.getUsers().subscribe((users: []) => {
       this.Users = users;
+      this.loader=false;
       console.log(this.Users);
     }, (err) => {
       console.log(err);
+      this.loader=false;
     });
   }
 
@@ -31,11 +36,26 @@ export class UsersListComponent implements OnInit {
   }
 
   deleteUser(userId: number) {
-    this.userService.deleteUser(userId).subscribe((res) => {
-      console.log(res);
-      this.getUsers();
-    }, (err) => {
-      console.log(err);
-    });
+    swal({
+      title: "Are you sure?",
+      text: "Once deleted, you will not be able to recover this record!",
+      icon: "warning",
+      buttons: true,
+      dangerMode: true,
+    })
+      .then((willDelete) => {
+        if (willDelete) {
+          this.userService.deleteUser(userId).subscribe((res) => {
+            swal(res['msg']['message'], {
+              icon: "success",
+            });
+            this.getUsers();
+          }, (err) => {
+            console.log(err);
+          });
+        } else {
+          swal("Your user record is safe!");
+        }
+      });
   }
 }
